MOODI-152 remove EnrichExecutor
